requirements.txt
torch
transformers
peft
datasets==3.6.0
trl
bitsandbytes
sentence-transformers
faiss-cpu
scikit-learn
accelerate
sacrebleu